+++ ui/temboardui/model/queries/instances-copy-as-csv.sql
@@ -1,22 +1,32 @@
 COPY (
   WITH
 		inventory AS (SELECT DISTINCT
-			i.hostname AS "Hostname",
-			i.pg_port AS "Port",
-			i.discover->'postgres'->'data_directory'#>>'{}' AS "PGDATA",
-			i.discover->'postgres'->'version_summary'#>>'{}' AS "Version",
-			e.name AS "Environment",
-			i.agent_address AS "Agent Address",
-			i.agent_port AS "Agent Port",
-			string_agg(DISTINCT plugin_name, ',') AS "Plugins",
-			i.comment AS "Comment"
-		FROM application.instances AS i
-		JOIN application.environments AS e ON e.id = i.environment_id
-		LEFT OUTER JOIN application.plugins
-				ON plugins.agent_address = i.agent_address
-				AND plugins.agent_port = i.agent_port
-		GROUP BY 1, 2, 3, 4, 5, 6, 7
-		ORDER BY 5, 1, 2
+      i.hostname AS "Hostname",
+      port AS "Port",
+      discover->'postgres'->'data_directory'#>>'{}' AS "PGDATA",
+      discover->'postgres'->'version_summary'#>>'{}' AS "Version",
+      e.name AS "Environment",
+      app.agent_address AS "Agent Address",
+      app.agent_port AS "Agent Port",
+      string_agg(DISTINCT plugins.plugin_name, ',') AS "Plugins",
+      app.comment AS "Comment",
+      string_agg(DISTINCT db.dbname, ',') AS "Databases"
+    FROM monitoring.hosts AS i
+    LEFT OUTER JOIN monitoring.instances AS monit
+         ON monit.host_id = i.host_id
+    LEFT OUTER JOIN application.instances AS app
+         ON app.hostname = i.hostname
+    LEFT OUTER JOIN application.plugins
+		     ON plugins.agent_address = app.agent_address
+		     AND plugins.agent_port = app.agent_port
+    LEFT OUTER JOIN application.environments AS e
+         ON e.id = app.environment_id
+    LEFT OUTER JOIN monitoring.metric_db_size_current AS db
+         ON db.instance_id = monit.instance_id
+         AND db.dbname NOT IN ('postgres', 'template1')
+         AND datetime > (now() - interval '5 minutes')
+GROUP BY 1, 2, 3, 4, 5, 6, 7
+ORDER BY 5, 1, 2
 	)
   SELECT * FROM inventory
 	WHERE concat_ws(